requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
email-validator
